+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This project tries to adhere to [Semantic Versioning](http://semver.org/).
 
 
 ## Unreleased
+### Added
+- Pass `git.path` setting to ungit, see ([#59](https://github.com/Hirse/vscode-ungit/issues/63))
+
 ### Changed
 - Updated Ungit to 1.4.43